refact: created some classes. also renamed.
