+++ arguments/__init__.py
@@ -1,7 +1,5 @@
-from arguments.ArgumentParser import ArgumentParser
-from arguments.util import parse_stdin
+from arguments.parse import Parser
 
 __all__ = [
-    "ArgumentParser",
-    "parse_stdin",
+    "Parser",
 ]
